Break out light classes into their own packages
