+++ pydeconz/models/light/configuration_tool.py
@@ -1,0 +1,9 @@
+"""Python library to connect deCONZ and Home Assistant to work together."""
+
+from . import DeconzLight
+
+
+class ConfigurationTool(DeconzLight):
+    """deCONZ hardware antenna."""
+
+    ZHATYPE = ("Configuration tool",)
